refactor: 레이싱 게임을 전반을 관리하는 도메인 객체 추가
