+++ src/main/kotlin/racingcar/model/RacingStatusManager.kt
@@ -7,18 +7,13 @@ class RacingStatusManager(
     private val getNumber: () -> Int,
     private val isMovable: (Int) -> Boolean,
 ) {
-    private var currentRacingStatus: List<Car> = cars
+    var currentRacingStatus: List<Car> = cars
+        private set
 
     fun getRacingStatus(): List<Car> {
         currentRacingStatus.forEach {
             if (isMovable(getNumber())) it.moveStep()
         }
-
         return currentRacingStatus
-    }
-
-    fun getWinners(): List<String> {
-        val maxPosition = currentRacingStatus.maxBy { it.position }.position
-        return currentRacingStatus.filter { it.position == maxPosition }.map { it.name }
     }
 }
